added hero image, javascript, and model
