+++ package.json
@@ -27,6 +27,7 @@
     "promise": "^7.1.1",
     "request": "^2.69.0",
     "serve-favicon": "~2.3.0",
+    "simple-oauth2": "^0.6.0",
     "underscore": "^1.8.3",
     "when": "^3.7.7"
   }
