+++ views/footer.html
@@ -1,6 +1,10 @@
 <!-- <div class='container footer'>
 		<p class='text-container'> [FOOTER] </p>
 	</div> -->
-	
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.0/jquery.min.js"></script>
+<script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>
+<script type="text/javascript" src="/javascripts/trans.js"></script>
+
+
 </body>
 </html>
