+++ views/header.html
@@ -14,8 +14,5 @@
 
 		<link rel='stylesheet' type='text/css' rel='stylesheet' href='/stylesheets/style.css'>
 
-	  	<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.0/jquery.min.js"></script>
-	  	<script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>
-
 	</head>
 	<body>
